added deleted

+++ webapp/app_psql/Routes/recipie.js
@@ -302,7 +302,8 @@ router.get('/recipie/:id', (req, res) => {
                                 }
                             })
                             .then(imageInformation => {
-                                res.header("Content-Type", 'application/json');
+                                if(imageInformation.length>0){
+                                    res.header("Content-Type", 'application/json');
 
                                 res.status(200).send(JSON.stringify(
 
@@ -332,6 +333,37 @@ router.get('/recipie/:id', (req, res) => {
                                         }
                                     }
                                 ));
+                                }
+                                else {
+                                    res.header("Content-Type", 'application/json');
+
+                                res.status(200).send(JSON.stringify(
+
+                                    {
+                                        "image":"NO IMAGE PRESENT",
+                                        "id": data[0].id,
+                                        "created_ts": data[0].created_date,
+                                        "updated_ts": data[0].updated_date,
+                                        "author_id": data[0].author_id,
+                                        "cook_time_in_min": data[0].cook_time_in_min,
+                                        "prep_time_in_min": data[0].prep_time_in_min,
+                                        "total_time_in_min": data[0].total_time_in_min,
+                                        "title": data[0].title,
+                                        "cusine": data[0].cusine,
+                                        "servings": data[0].servings,
+                                        "ingredients": data[0].ingredients,
+                                        "steps": data[0].steps,
+                                        "nutrition_information": {
+                                            "calories": nutrition_information[0].calories,
+                                            "cholesterol_in_mg": nutrition_information[0].cholesterol_in_mg,
+                                            "sodium_in_mg": nutrition_information[0].sodium_in_mg,
+                                            "carbohydrates_in_grams": nutrition_information[0].carbohydrates_in_grams,
+                                            "protein_in_grams": nutrition_information[0].protein_in_grams
+                                        }
+                                    }
+                                ));
+                                }
+                                
                             })
                     })
             }
@@ -584,17 +616,21 @@ router.delete('/recipie/:id/image/:imageId', (req, res) => {
                         });
                     } else if (result) {
 
-                        if(){
-                            // recipe invalid
-                        }
-                        if() {
-                            // image mapping not present
-                        }
-                        else {
-                            //delete
-                        }
+                        db.image.findAll({
+                            where: {
+                                recipe_id: req.params.id
+                            }
+                        })
+                        .then(image_data => {
+                            if(image_data.length>0){
 
-                        db.image.destroy({
+                            if(image_data[0].image_id!=req.params.imageId){
+                                res.status(404).json({
+                                    message:"No Image Id Found"
+                                })
+                            }
+                            else {
+                                db.image.destroy({
                                 where: {
                                     recipe_id: req.params.id
                                 }
@@ -602,7 +638,7 @@ router.delete('/recipie/:id/image/:imageId', (req, res) => {
                             .then(deletedImage => {
                                 if (deletedImage > 0) {
                                     res.status(200).json({
-                                        deletedRecipe
+                                        deletedImage
                                     })
                                 } else {
                                     res.status(404).json({
@@ -610,10 +646,21 @@ router.delete('/recipie/:id/image/:imageId', (req, res) => {
                                     })
                                 }
                             })
-                            .catch(err => res.status(406).json({
-                                message: err.message
-                            }));
-
+                            }
+                        }
+                        else {
+                            //if(image_data[0].recipe_id!=req.params.id){
+                                res.status(404).json({
+                                    message:"No Recipe ID found"
+                                })
+                            //}
+                            
+                        }
+                        })
+                        .catch(err => res.status(406).json({
+                            message: err.message,
+                            //message: "No recipe ID found"
+                        }));
 
                     } else {
                         res.status(401).json({
